% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeHG19Promoters.R
\name{makeHG19Promoters}
\alias{makeHG19Promoters}
\title{Generate promoter annotation of hg19 refseq genes}
\usage{
makeHG19Promoters(up = 2000, down = 2000)
}
\arguments{
\item{up}{An integer value defining the number of nucleotides toward the 5' end relative to
the transcription start site.}

\item{down}{An integer value defining the number of nucleotides toward the 3' end relative to
the transcription start site.}
}
\value{
A GRanges object containing promoter annotation.
}
\description{
Generate promoter annotation of hg19 refseq genes
}
\examples{
makeHG19Promoters(2000, 2000)
makeHG19Promoters(1000, 1000)
}
