% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promoterDHcR.R
\name{promoterDHcR}
\alias{promoterDHcR}
\title{Calculate promoter DHcR}
\usage{
promoterDHcR(file_name, pro, min_cpgs = 5)
}
\arguments{
\item{file_name}{A tab-separated values input file. The input file contains details of differentially methylated
cytosines (DMCs) with following columns (V1 to V11): chr, pos, numC in control, numC + numT in control,
numC in tumor, numC + numT in tumor, CpG methylation ratio (tumor methylation / control methylation),
chi-squared test p-value, adjusted p-value, significance, hyper or hypo in tumor.
Details of generating this type of files were described in Pan et al, Cancer Systems Biology, 2018.}

\item{pro}{A GRanges object containing promoter annotation.}

\item{min_cpgs}{An integer value defining minimum number of covered CpGs for each promoter.}
}
\value{
A data frame summarizing promoter DHcR levels.
    \itemize{
      \item Hugo: Hugo symbol
      \item Depth: Average sequencing depth
      \item HCs: Number of HCs
      \item CpGs: Number of covered CpGs
      \item DHcR: DHcR level
    }
}
\description{
Promoter differentially hypermethylated cytosine ratio (DHcR) was defined as the ratio of
    hypermethylated cytosines (HCs) to the total number of promoter CpGs profiled.
}
\examples{
promoterDHcR(system.file("extdata", "DMC.SRR2069925.txt", package = "MethSig"), pro=pro)

}
