% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promoterPDR.R
\name{promoterPDR}
\alias{promoterPDR}
\title{Calculate promoter PDR}
\usage{
promoterPDR(file_name, pro, min_cpgs = 3)
}
\arguments{
\item{file_name}{A tab-separated values input file. The input file contains PDR of single CpG with
following columns: chr, start, strand, ConMethReadCount, ConUMethReadCount, DisReadCount, NAReadCount.
Details of PDR were described in Landau et al., Cancer Cell, 2014. extdata/pdrCall_from_Bismark.py
can be used to call PDR of single CpG from Bismark outputs (files starting with CpG_OB or CpG OT).}

\item{pro}{A GRanges object containing promoter annotation.}

\item{min_cpgs}{Single integer value >= 0L. The minimum number available CpGs for each promoter.}
}
\value{
A data frame summarizing promoter PDR levels.
    \itemize{
      \item Hugo: Hugo symbol
      \item PDR: PDR level
    }
}
\description{
Promoter proportion of discordant reads (PDR) was defined as the average PDR of all the CpGs inside.
}
\details{
If all the CpGs on a specific read are methylated, or all of the CpGs on a read are unmethylated,
    the read is classified as concordant; otherwise it is classified as discordant. At each CpG,
    the PDR is equal to the number of discordant reads that cover that location divided by
    the total number of read that cover that location. Only CpGs with read depth greater than 10 reads
    and covered by reads that contain at least 4 CpGs were included into the analysis.
}
\examples{
promoterDHcR('PDR.SRR2069925.txt', pro)

}
