% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvalueBetaReg.R
\name{pvalueBetaReg}
\alias{pvalueBetaReg}
\title{Infer sample-specific hypermethylation}
\usage{
pvalueBetaReg(formula, data)
}
\arguments{
\item{formula}{An object of class "formula" (or one that can be coerced to that class):
a symbolic description of the model to be fitted. The name of y variable must DHcR_Tumor_Beta.}

\item{data}{A data frame object generated by makeInputMatrix function (input matrix).
The object contains following columns:
Id, Hugo, Covariates included in matCV, PDR_Tumor, Depth_Tumor, CpGs_Tumor and DHcR_Tumor.}
}
\value{
A data frame summarizing information of sample-specific hypermethylation inference. Addtional columns
    were added to input matrix.
    \itemize{
      \item DHcR_Tumor_Beta: Transformed data of DHcR_Tumor (from [0, 1] to (0, 1))
      \item Beta_Response: Response value from beta-regression
      \item Beta_Precision: Precisison value from beta-regression
      \item Pvalue: P-value of sample-specific hypermethylation inference
    }
}
\description{
Expected DHcR was estimated through beta-regression first. Next, expected DHcR was tested
    against to observed DHcR to evalute if there was promoter hypermethylation.
}
\examples{
pvalueBetaReg('DHcR_Tumor_Beta~DHcR_Normal+PDR_Normal+GEXP_Normal+Reptime+PDR_Tumor+Depth_Tumor+CpGs_Tumor',
    inputMat)

}
