% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvalueCombine.R
\name{pvalueCombine}
\alias{pvalueCombine}
\title{Infer cohort-prevalent hypermethylation}
\usage{
pvalueCombine(data, K = 10, times = 100, r = 2)
}
\arguments{
\item{data}{A data frame object generated by \emph{pvalueBetaReg} containing following columns:
Id, Hugo, Covariates included in \emph{matCV}, PDR_Tumor, Depth_Tumor, CpGs_Tumor, DHcR_Tumor,
DHcR_Tumor_Beta, Beta_Response, Beta_Precision and Pvalue.
An example can be loaded by using \emph{invisible(pvalByGenePt)}.}

\item{K}{An integer value defining the number of selected samples in random sampling.}

\item{times}{An integer value defining the iterative times in random sampling.}

\item{r}{Use the rth smallest p-value in Wilkinson combination.}
}
\value{
A data frame summarizing cohort-prevalent hypermethylation inference.
    \itemize{
      \item Hugo: Hugo symbol
      \item Rank: Rank of the promoter by p-value
      \item Sample_Size: Number of samples evaluated
      \item Pvalue: Lower quartile of Wilkinson combined p-values from random sampling
      \item Padjust: Benjamini-Hochberg adjusted p-value
    }
}
\description{
Determine if promoter hypermethylation is over-represented in the cohort.
}
\examples{
pvalueCombine(data = invisible(pvalByGenePt))

}
