% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeInputMatrix.R
\name{makeInputMatrix}
\alias{makeInputMatrix}
\title{Make input matrix}
\usage{
makeInputMatrix(names_list, matCV, pro, input_dir, min_cpgs = c(5, 3))
}
\arguments{
\item{names_list}{A list of sample names.}

\item{matCV}{A matrix of covariates which are included into the analysis. Ideally, users can include any covariate
that will be correlated with promoter DHcR in tumor samples and a hugo symbol column needs to be included.
An example can be loaded by using \emph{invisible(matCV)}. This covariate matrix contains following columns:
Hugo, DHcR_Normal, PDR_Normal, GEXP_Normal, Reptime.}

\item{pro}{A GRanges object containing promoter annotation.}

\item{input_dir}{A string indicating where input files used to calculate DHcR and PDR of all the samples
are stored.
Input files used to calculate DHcR should be named as DMC.sample_name.txt format.
Input files used to calculate PDR should be named as PDR.sample_name.txt format.}

\item{min_cpgs}{A list of two integer values: the minimum number of CpGs requied by \emph{promoterDHcR} and
the minimum number of CpGs required by \emph{promoterPDR}.}
}
\value{
A data frame used as input for sample-specific hypermethylation inference.
    \itemize{
      \item Id: Sample id
      \item Hugo: Hugo symbol
      \item Covariates included in matCV
      \item PDR_Tumor: PDR level in tumor samples
      \item Depth_Tumor: Average sequencing depth in tumor samples
      \item CpGs_Tumor: Number of CpGs covered in tumor samples
      \item DHcR_Tumor: DHcR level in tumor samples
    }
}
\description{
Add DHcR, PDR and sequencing depth information of samples to be inferred into input matrix.
}
\examples{
makeInputMatrix(names_list = as.list('SRR2069925'),
                matCV = invisible(matCV),
                pro = makeHG19Promoters(),
                input_dir = system.file("extdata", "", package = "MethSig"))

}
