% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promoterPDR.R
\name{promoterPDR}
\alias{promoterPDR}
\title{Calculate promoter PDR}
\usage{
promoterPDR(file_name, pro, min_cpgs = 3)
}
\arguments{
\item{file_name}{A tab-separated values input file. The input file contains details of single CpG PDR with
following columns: chr, start, strand, ConMethReadCount, ConUMethReadCount, DisReadCount, NAReadCount.
Details of PDR were described in Landau \emph{et al.}, \emph{Cancer Cell}, 2014.
extdata/pdrCall_from_Bismark.py can be used to call PDR of single CpG from
Bismark (Krueger \emph{et al.}, \emph{Bioinformatics}, 2011) outputs (files starting with CpG_OB or CpG_OT).}

\item{pro}{A GRanges object containing promoter annotation.}

\item{min_cpgs}{An integer value defining the minimum number of available CpGs for each promoter.}
}
\value{
A data frame summarizing promoter PDR levels.
    \itemize{
      \item Hugo: Hugo symbol
      \item PDR: PDR level
    }
}
\description{
Promoter proportion of discordant reads (PDR) is defined as the average PDR of all the CpGs inside.
}
\examples{
promoterPDR(file_name = system.file("extdata", "PDR.SRR2069925.txt", package = "MethSig"),
            pro = makeHG19Promoters())

}
