% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvalueBetaReg.R
\name{pvalueBetaReg}
\alias{pvalueBetaReg}
\title{Infer sample-specific hypermethylation}
\usage{
pvalueBetaReg(formula, data)
}
\arguments{
\item{formula}{An object of class "formula" (or one that can be coerced to that class):
a symbolic description of the model to be fitted. The name of y variable should be DHcR_Tumor_Beta.}

\item{data}{A data frame object generated by \emph{makeInputMatrix}.
An example can be loaded by using \emph{invisible(inputMat)}.
The object contains following columns:
Id, Hugo, Covariates included in matCV, PDR_Tumor, Depth_Tumor, CpGs_Tumor and DHcR_Tumor.}
}
\value{
A data frame summarizing information of sample-specific hypermethylation inference. Addtional columns
    are added to the input matrix.
    \itemize{
      \item DHcR_Tumor_Beta: Transformed data of DHcR_Tumor (from [0, 1] to (0, 1))
      \item Beta_Response: Response value estimated by beta regression
      \item Beta_Precision: Precision value estimated by beta regression
      \item Pvalue: P-value of sample-specific hypermethylation inference
    }
}
\description{
Expected DHcR is estimated through beta regression model first. Next, expected DHcR is tested
    against observed DHcR to evalute if there is promoter hypermethylation.
}
\examples{
pvalueBetaReg(formula = 'DHcR_Tumor_Beta~DHcR_Normal+PDR_Normal+GEXP_Normal+Reptime+PDR_Tumor+Depth_Tumor+CpGs_Tumor',
              data = invisible(inputMat))

}
