% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvalueCombine.R
\name{pvalueCombine}
\alias{pvalueCombine}
\title{Infer cohort-prevalent hypermethylation}
\usage{
pvalueCombine(data, K = 10, times = 100, r = 2)
}
\arguments{
\item{data}{A data frame object generated by pvalueBetaReg function. The object contains following columns:
Id, Hugo, Covariates included in matCV, PDR_Tumor, Depth_Tumor, CpGs_Tumor, DHcR_Tumor, DHcR_Tumor_Beta,
Beta_Response, Beta_Precision and Pvalue.}

\item{K}{Single integer value > 0L. The size of random sampling.}

\item{times}{Single integer value > 0L. The times of random sampling.}

\item{r}{Use the rth smallest p-value in Wilkinson combination.}
}
\value{
A data frame summarizing inference of cohort-prevalent hypermethylation.
    \itemize{
      \item Hugo: Hugo symbol
      \item Rank: Rank of the promoter by p-value
      \item Sample_Size: Number of samples evaluated in this promoter
      \item Pvalue: Lower quartile of wilkinson combined p-value from random sampling
      \item Padjust: Benjamini-Hochberg adjusted p-value
    }
}
\description{
Determine if promoter hypermethylation is over-represented in the cohort.
}
\examples{
pvalueCombine(pvalByGenePt)

}
